# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BlastArena API
  version: 1.0.0
  description: |
    REST API for BlastArena, a multiplayer online grid-based explosive arena game.
    All endpoints are mounted under the `/api` prefix.
    Authentication uses JWT Bearer tokens (access token) and httpOnly cookies (refresh token).
servers:
  - url: /api
    description: API base path
...
security:
  - BearerAuth: []
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using cookie
      description: Uses the httpOnly refreshToken cookie. No Authorization header needed.
      security:
        - CookieAuth: []
      operationId: refreshToken
      responses:
        '200':
          description: Token refreshed
          headers:
            Set-Cookie:
              description: Updated httpOnly refreshToken cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: No refresh token or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /user/profile:
    get:
      tags: [User]
      summary: Get current user profile
      operationId: getUserProfile
      responses:
        '200':
          description: User profile with stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [User]
      summary: Update user profile
      operationId: updateUserProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 20
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Username taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/email:
    post:
      tags: [User]
      summary: Change email address
      description: Admins get direct update; regular users receive a confirmation email.
      operationId: changeEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
      responses:
        '200':
          description: Email updated or confirmation sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [User]
      summary: Cancel pending email change
      operationId: cancelEmailChange
      responses:
        '200':
          description: Pending change cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user/password:
    post:
      tags: [User]
      summary: Change password
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                  maxLength: 128
                newPassword:
                  type: string
                  minLength: 8
                  maxLength: 128
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Validation error or same password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user/confirm-email/{token}:
    get:
      tags: [User]
      summary: Confirm email change
      description: The link sent by email. On success the browser is redirected to `{APP_URL}?emailChanged=true`, like `/auth/verify-email`.
      security: []
      operationId: confirmEmailChange
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Email updated; redirect to the app with `?emailChanged=true`
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ===================== Lobby =====================
...
  /user/privacy:
    put:
      tags: [User]
      summary: Update privacy settings
      operationId: updatePrivacy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivacySettings'
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ===================== User — Language =====================
  /user/language:
    put:
      tags: [User]
      summary: Update preferred language
      operationId: updateLanguage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [language]
              properties:
                language:
                  $ref: '#/components/schemas/SupportedLanguage'
      responses:
        '200':
          description: Language updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  language:
                    $ref: '#/components/schemas/SupportedLanguage'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ===================== User — Buddy Settings =====================
  /user/buddy-settings:
    get:
      tags: [User]
      summary: Get buddy mode settings
      operationId: getBuddySettings
      responses:
        '200':
          description: Current buddy settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuddySettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [User]
      summary: Update buddy mode settings
      operationId: updateBuddySettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuddySettings'
      responses:
        '200':
          description: Updated buddy settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuddySettings'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ===================== User — Delete Account =====================
  /user/account:
    delete:
      tags: [User]
      summary: Delete own account
      description: Hard-deletes the account after password confirmation. Admin accounts cannot self-delete.
      operationId: deleteAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Wrong password or admin account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ===================== Leaderboard =====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token passed in Authorization header
    CookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: httpOnly refresh token cookie (set by login/refresh endpoints)
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Message:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ---- Enums ----
    UserRole:
      type: string
      enum: [user, moderator, admin]
...
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/PublicUser'
        accessToken:
          type: string
        expiresIn:
          type: integer
          description: Token expiry in seconds
    PublicUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'

    # ---- User ----
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        emailVerified:
          type: boolean
        role:
          $ref: '#/components/schemas/UserRole'
        lastLogin:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        totalMatches:
          type: integer
        totalWins:
          type: integer
        totalKills:
          type: integer
        totalDeaths:
          type: integer

    # ---- Lobby ----
...
    SupportedLanguage:
      type: string
      enum: [en, de, fr, es, it, pt, pl, nl, tr, sv, nb, da]
...
    PrivacySettings:
      type: object
      properties:
        isProfilePublic:
          type: boolean
        acceptFriendRequests:
          type: boolean
    BuddySettings:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 20
        color:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
          description: Hex color code
        size:
          type: number
          minimum: 0.4
          maximum: 0.8

    # ---- Leaderboard ----
...